… exporter/splunkhecexporter/go.sum
… go.sum
… receiver/splunkhecreceiver/go.sum