**Description:**
Use jsoniter instead of std json lib to marshal/unmarshal json with hec.

**Testing:**

Exporter:
Adding a performance test that shows an improvement of ~ 60% in performance.

Before:
```
goos: darwin
goarch: amd64
pkg: github.com/open-telemetry/opentelemetry-collector-contrib/exporter/splunkhecexporter
cpu: Intel(R) Core(TM) i9-9980HK CPU @ 2.40GHz
BenchmarkPushLogRecords
BenchmarkPushLogRecords-16    	  463678	      2243 ns/op
```

After:
```
goos: darwin
goarch: amd64
pkg: github.com/open-telemetry/opentelemetry-collector-contrib/exporter/splunkhecexporter
cpu: Intel(R) Core(TM) i9-9980HK CPU @ 2.40GHz
BenchmarkPushLogRecords
BenchmarkPushLogRecords-16    	  747003	      1592 ns/op
```

Receiver:
Adding a performance test that shows an improvement of ~50% in performance.

Before:
```
goos: darwin
goarch: amd64
pkg: github.com/open-telemetry/opentelemetry-collector-contrib/receiver/splunkhecreceiver
cpu: Intel(R) Core(TM) i9-9980HK CPU @ 2.40GHz
BenchmarkHandleReq
BenchmarkHandleReq-16    	    1192	    879736 ns/op
```

After:
```
goos: darwin
goarch: amd64
pkg: github.com/open-telemetry/opentelemetry-collector-contrib/receiver/splunkhecreceiver
cpu: Intel(R) Core(TM) i9-9980HK CPU @ 2.40GHz
BenchmarkHandleReq
BenchmarkHandleReq-16    	    1780	    642720 ns/op
```