@dmitryax 2021-12-05
Should we have a benchmark test for this change to see how it performs specifically with hec payload?