@dmitryax 2021-12-06
It should test `encodeBodyEvents` function I believe, so it'll go to client_test.go