@atoulme 2021-12-06
Thanks! I added the benchmark, please take a look. I hope I have my math right :)